requirements.txt
pytube==15.0.0
certifi==2025.1.31
spotipy>=2.23.0
python-dotenv>=1.0.0 